# feat: Speed up symbol parsing by minimizing allocations