--- a/bindings/go/scip/assertions.go
+++ b/bindings/go/scip/assertions.go
@@ -1,0 +1,9 @@
+//go:build asserts
+
+package scip
+
+func assert(cond bool, msg string) {
+	if !cond {
+		panic(msg)
+	}
+}