--- a/bindings/go/scip/assertions_noop.go
+++ b/bindings/go/scip/assertions_noop.go
@@ -1,0 +1,6 @@
+//go:build !asserts
+
+package scip
+
+// assert is a noop in release builds - the implementation is in assertions.go
+func assert(cond bool, msg string) {}