--- a/bindings/go/scip/symbol_formatter.go
+++ b/bindings/go/scip/symbol_formatter.go
@@ -2,6 +2,8 @@ package scip
 
 import (
 	"strings"
+
+	"github.com/sourcegraph/scip/bindings/go/scip/internal/shared"
 )
 
 // SymbolFormatter configures how to format an SCIP symbol.
@@ -139,7 +141,7 @@ func writeEscapedPackage(b *strings.Builder, name string) {
 func writeSuffixedDescriptor(b *strings.Builder, identifier string, suffixes ...rune) {
 	escape := false
 	for _, ch := range identifier {
-		if !isSimpleIdentifierCharacter(ch) {
+		if !shared.IsSimpleIdentifierCharacter(ch) {
 			escape = true
 			break
 		}