--- a/dev/sample_indexes/.gitignore
+++ b/dev/sample_indexes/.gitignore
@@ -1,0 +1,3 @@
+# Ignore any SCIP indexes as well as any files with statistics data.
+*.scip
+*.json