… .github/workflows/golang.yml
… Development.md
… bindings/go/scip/internal/compat_test.go
… bindings/go/scip/memtest/low_mem_test.go
… bindings/go/scip/sort_test.go
… bindings/go/scip/symbol_parser.go
… bindings/go/scip/symbol_test.go
… bindings/haskell/src/Proto/Scip.hs
… bindings/rust/src/generated/scip.rs
… dev/sample_indexes/indexes-metadata.json
… docs/scip.md
… go.mod
… go.sum
… scip.proto