NOTE: The diff looks large, but the majority of that is because of scheme
changes due to a small comment that I added.

## What

Changes the symbol parsing logic to minimize allocations. In particular,
when we only care about _validating_ symbols (e.g. during document
canonicalization when ingesting uploads), there is really ~no need to
allocate any strings at all. Validation and parsing share most of the
underlying code -- the only change is we create "writer" types which
will discard writes (and hence any internal buffer growth) when we're
only in validation mode.

## Why

Ideally, we want to validate all symbols that we enter into the DB,
(and we also want to have fast splitting of symbols) so it's valuable
to have the overhead be as low as possible. In the validation case,
we only make minimal heap allocations in the error case (there is a
test which makes sure we don't allocate in the non error cases).

## Benchmarks

I ran some benchmarks with sample SCIP indexes located here: (Sourcegraph-internal)
https://drive.google.com/drive/folders/1z62Se7eHaa5T89a16-y7s0Z1qbRY4VCg

Once the indexes are decompressed into `dev/sample_indexes`, you can run

```
# Benchmarks
go run ./bindings/go/scip/speedtest
# Compatibility test with old parser
go test ./... -run TestParseCompat -tags asserts
```

Symbol parse (v1) represents the older symbol parsing logic;
Symbol parse (v2) represents the newer symbol parsing logic.
I also added a validation helper function on top of the newer parser;
that is also benchmarked separately.

symbol parse (v2) is noticeably slower than validation because of
allocations needed so that we can have symbol parse (v1) and (v2)
-- the old parsing logic would always return a new `*scip.Symbol`,
so it'd be an unfair comparison if we just pre-allocated everything
for benchmarking (v2). It is possible to get symbol parse (v2) to validation level
speed by pre-allocating arrays of `scip.Package` and `scip.Descriptor`
values up-front and essentially passing pointers to those inside those to
successive calls to `ParseSymbolUTF8With`.

```
Benchmark for "/Users/varun/Code/scip/dev/sample_indexes/django-1.scip":
+-------------------------------+------------+-------+------------+-------+----------+-------+
| BENCHMARK / ITERATIONS        | 1000       | RATIO | 10000      | RATIO | 100000   | RATIO |
+-------------------------------+------------+-------+------------+-------+----------+-------+
| Symbol parse (v1) - Speed     | 1.074µs/op | -     | 1.059µs/op | -     | 928ns/op | -     |
| Symbol parse (v2) - Speed     | 528ns/op   | 0.49x | 620ns/op   | 0.59x | 620ns/op | 0.67x |
| Symbol validate (v2) - Speed  | 359ns/op   | 0.33x | 421ns/op   | 0.40x | 413ns/op | 0.45x |
| Symbol parse (v1) - Allocs    | 101535B/op | -     | 10153B/op  | -     | 1015B/op | -     |
| Symbol parse (v2) - Allocs    | 41218B/op  | 0.41x | 4121B/op   | 0.41x | 412B/op  | 0.41x |
| Symbol validate (v2) - Allocs | 4B/op      | 0.00x | 0B/op      | 0.00x | 0B/op    | 0.00x |
+-------------------------------+------------+-------+------------+-------+----------+-------+
Benchmark for "/Users/varun/Code/scip/dev/sample_indexes/flink-1.scip":
+-------------------------------+------------+-------+------------+-------+----------+-------+
| BENCHMARK / ITERATIONS        | 1000       | RATIO | 10000      | RATIO | 100000   | RATIO |
+-------------------------------+------------+-------+------------+-------+----------+-------+
| Symbol parse (v1) - Speed     | 1.334µs/op | -     | 1.216µs/op | -     | 962ns/op | -     |
| Symbol parse (v2) - Speed     | 665ns/op   | 0.50x | 686ns/op   | 0.56x | 694ns/op | 0.72x |
| Symbol validate (v2) - Speed  | 397ns/op   | 0.30x | 392ns/op   | 0.32x | 400ns/op | 0.42x |
| Symbol parse (v1) - Allocs    | 107481B/op | -     | 10748B/op  | -     | 1074B/op | -     |
| Symbol parse (v2) - Allocs    | 61517B/op  | 0.57x | 6151B/op   | 0.57x | 615B/op  | 0.57x |
| Symbol validate (v2) - Allocs | 0B/op      | 0.00x | 0B/op      | 0.00x | 0B/op    | 0.00x |
+-------------------------------+------------+-------+------------+-------+----------+-------+
Benchmark for "/Users/varun/Code/scip/dev/sample_indexes/llvm-project-1.scip":
+-------------------------------+-----------+-------+----------+-------+----------+-------+
| BENCHMARK / ITERATIONS        | 1000      | RATIO | 10000    | RATIO | 100000   | RATIO |
+-------------------------------+-----------+-------+----------+-------+----------+-------+
| Symbol parse (v1) - Speed     | 941ns/op  | -     | 651ns/op | -     | 516ns/op | -     |
| Symbol parse (v2) - Speed     | 387ns/op  | 0.41x | 442ns/op | 0.68x | 363ns/op | 0.70x |
| Symbol validate (v2) - Speed  | 229ns/op  | 0.24x | 230ns/op | 0.35x | 174ns/op | 0.34x |
| Symbol parse (v1) - Allocs    | 49424B/op | -     | 4942B/op | -     | 494B/op  | -     |
| Symbol parse (v2) - Allocs    | 33525B/op | 0.68x | 3352B/op | 0.68x | 335B/op  | 0.68x |
| Symbol validate (v2) - Allocs | 0B/op     | 0.00x | 0B/op    | 0.00x | 0B/op    | 0.00x |
+-------------------------------+-----------+-------+----------+-------+----------+-------+
Benchmark for "/Users/varun/Code/scip/dev/sample_indexes/rust-1.scip":
+-------------------------------+-----------+-------+----------+-------+----------+-------+
| BENCHMARK / ITERATIONS        | 1000      | RATIO | 10000    | RATIO | 100000   | RATIO |
+-------------------------------+-----------+-------+----------+-------+----------+-------+
| Symbol parse (v1) - Speed     | 664ns/op  | -     | 636ns/op | -     | 739ns/op | -     |
| Symbol parse (v2) - Speed     | 450ns/op  | 0.68x | 443ns/op | 0.70x | 498ns/op | 0.67x |
| Symbol validate (v2) - Speed  | 278ns/op  | 0.42x | 288ns/op | 0.45x | 301ns/op | 0.41x |
| Symbol parse (v1) - Allocs    | 70200B/op | -     | 7020B/op | -     | 702B/op  | -     |
| Symbol parse (v2) - Allocs    | 38142B/op | 0.54x | 3814B/op | 0.54x | 381B/op  | 0.54x |
| Symbol validate (v2) - Allocs | 0B/op     | 0.00x | 0B/op    | 0.00x | 0B/op    | 0.00x |
+-------------------------------+-----------+-------+----------+-------+----------+-------+
Benchmark for "/Users/varun/Code/scip/dev/sample_indexes/shopify-api-ruby-1.scip":
+-------------------------------+------------+-------+----------+-------+----------+-------+
| BENCHMARK / ITERATIONS        | 1000       | RATIO | 10000    | RATIO | 100000   | RATIO |
+-------------------------------+------------+-------+----------+-------+----------+-------+
| Symbol parse (v1) - Speed     | 1.024µs/op | -     | 914ns/op | -     | 893ns/op | -     |
| Symbol parse (v2) - Speed     | 5.607µs/op | 5.48x | 349ns/op | 0.38x | 340ns/op | 0.38x | (*)
| Symbol validate (v2) - Speed  | 236ns/op   | 0.23x | 203ns/op | 0.22x | 203ns/op | 0.23x |
| Symbol parse (v1) - Allocs    | 65794B/op  | -     | 6579B/op | -     | 657B/op  | -     |
| Symbol parse (v2) - Allocs    | 30391B/op  | 0.46x | 3039B/op | 0.46x | 303B/op  | 0.46x |
| Symbol validate (v2) - Allocs | 0B/op      | 0.00x | 0B/op    | 0.00x | 0B/op    | 0.00x |
+-------------------------------+------------+-------+----------+-------+----------+-------+
Benchmark for "/Users/varun/Code/scip/dev/sample_indexes/typescript-1.scip":
+-------------------------------+-----------+-------+----------+-------+----------+-------+
| BENCHMARK / ITERATIONS        | 1000      | RATIO | 10000    | RATIO | 100000   | RATIO |
+-------------------------------+-----------+-------+----------+-------+----------+-------+
| Symbol parse (v1) - Speed     | 555ns/op  | -     | 597ns/op | -     | 572ns/op | -     |
| Symbol parse (v2) - Speed     | 450ns/op  | 0.81x | 389ns/op | 0.65x | 385ns/op | 0.67x |
| Symbol validate (v2) - Speed  | 189ns/op  | 0.34x | 225ns/op | 0.38x | 222ns/op | 0.39x |
| Symbol parse (v1) - Allocs    | 62000B/op | -     | 6200B/op | -     | 620B/op  | -     |
| Symbol parse (v2) - Allocs    | 37137B/op | 0.60x | 3713B/op | 0.60x | 371B/op  | 0.60x |
| Symbol validate (v2) - Allocs | 0B/op     | 0.00x | 0B/op    | 0.00x | 0B/op    | 0.00x |
+-------------------------------+-----------+-------+----------+-------+----------+-------+
```

There is one surprising case where parsing is much slower when only testing against
1000 occurrences with the newer parser, which I've marked with a `(*)` --
that seems to be somewhat reproducible. However, I haven't spent time on investigating
that because it seems like the speed improvements are present at higher occurrence counts.

### Test plan

Added compatibility tests for the old parser vs the new parser and ran
that against a bunch of existing indexes.